requirements.txt
django==1.11.20
amqp==2.5.0
appdirs==1.4.3
attrs==19.1.0
backcall==0.1.0
backports.csv==1.0.7
beautifulsoup4==4.7.1
billiard==3.6.0.0
bioblend==0.12.0
biopython==1.73
bioservices==1.6.0
boto==2.49.0
cached-property==1.5.1
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
colorlog==4.0.2
cycler==0.10.0
decorator==4.4.0
deepdiff==4.0.6
defusedxml==0.6.0
django-autocomplete-light==3.2.10
django-bootstrap3==11.0.0
django-bulk-update==2.2.0
django-celery-results==1.1.2
django-filter==1.1.0
django-galaxy==0.0.1
django-gfiles==0.0.1
django-mbrowse==0.0.1
django-misa==0.0.1
django-mogi==0.0.1
django-rest-framework==0.1.0
django-sb-admin==0.3.7
django-tables2==1.21.0
django-tables2-column-shifter==0.5.2
djangorestframework==3.9.4
easydev==0.9.37
et-xmlfile==1.0.1
fs==2.2.1
future==0.17.1
gevent==1.4.0
greenlet==0.4.15
grequests==0.3.0
idna==2.8
ipython==7.5.0
ipython-genutils==0.2.0
isatools==0.10.3
iso8601==0.1.12
jdcal==1.4.1
jedi==0.13.3
Jinja2==2.10.1
jsonpickle==1.2
jsonschema==3.0.1
jupyter-core==4.4.0
kiwisolver==1.1.0
kombu==4.6.3
lxml==4.3.4
MarkupSafe==1.1.1
matplotlib==3.1.0
mysqlclient==1.3.14
mzml2isa==1.0.1
nbformat==4.4.0
networkx==2.3
numpy==1.16.4
odfpy==1.4.0
openpyxl==2.6.2
ordered-set==3.1.1
pandas==0.24.2
parso==0.4.0
pexpect==4.7.0
pickleshare==0.7.5
plotly==3.10.0
progressbar2==3.42.0
prompt-toolkit==2.0.9
pronto==0.12.2
ptyprocess==0.6.0
PubChemPy==1.0.4
Pygments==2.4.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2019.1
PyYAML==5.1.1
requests==2.22.0
requests-cache==0.5.0
requests-toolbelt==0.9.1
retrying==1.3.3
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
soupsieve==1.9.1
sqlparse==0.3.0
suds-jurko==0.6
tablib==0.13.0
traitlets==4.3.2
urllib3==1.25.3
vine==1.3.0
wcwidth==0.1.7
wrapt==1.11.2
xlrd==1.2.0
xlwt==1.3.0
xmltodict==0.12.0

